feat(kubernetes): update ingress configurations and PostgreSQL image version
This commit modifies the ingress configurations for Aithena and RabbitMQ WebSocket to include a specific host (`polus2.ncats.nih.gov`) and changes the path types to `ImplementationSpecific` for better routing flexibility. Additionally, the PostgreSQL image version is updated from `latest` to `17` in the Litellm database deployment for improved stability and performance. These changes enhance the deployment's routing capabilities and ensure a more reliable database environment.